Reduce Number of SwiftUI Redraws
- Make sub views equatable
- Make buttons always return true
- Remove nested observable object (was causing random issues)

Improves performance. Text inputs are the primary remaining bottleneck.

The end goal is freeing up the main thread enough that the image editing can be moved to a background thread with frequent main thread updates

Spent a lot of time in instruments to figure these changes out

There is a refactor of image editing on the horizon