Custom Preset Editor Undo/Redo
Moves the selection state into the store so it can be easily managed by undo and redo. Not the best place in theory but it works the best currently.

Also it looks like ventura fixed the side bar text field issue, so the workaround has been removed (the workaround was also causing a crash)